Bugfix 1 - update recipes
Modifies the recipeGetSuccess function to check the length of the
object. If length is 0 then display a message on the page to create
a recipe - adds back hidden class on the chance the page has not been
left. Otherwise, recipes are displayed and the update button is
made hidden again.